Clean up all boot files and split DescriptorSet module in 3
- Descriptor
- Descriptor.Pool
- Descriptor.Set
