+++ TODO
@@ -110,7 +110,7 @@
 [ ] Perhaps use more efficient list-like type by default for creating meshes
 [ ] Dear imgui specific render pipeline
 [x] Custom descriptor set #0 through Render Pipeline constructors + edit function similar to Materials
-[ ] Untangle World {-# SOURCE #-} imports by creating another module for it
+[x] Untangle World {-# SOURCE #-} imports by creating another module for it
 [ ] Hex map tutorial series https://catlikecoding.com/unity/tutorials/hex-map/
 [x] Compatible mesh constraint with working planets example
 [x] Get rid of VertexArray altogether (use Storable Vector directly)?
