+++ hie.yaml
@@ -1,3 +1,3 @@
 cradle:
   cabal:
-    component: "all"
+    component: "lib:ghengin-core"
